Finished most details regarding user profiles
A user is created in the MainActivity function immediately. The unique ID is taken from the firebase installation ID associated with the app. This should be persistent between runs of the app but will most likely change if the app has to be reinstalled. I am unsure of this currently. The user is added to the database if the ID is new and contact info is taken from the database if it isn't.

There is now a function in in the main activity to switch to the DisplayUserProfile activity. This sends the created user over. When it is finished, it updates the current user.

The update profile button now updates the database and stores the changes into the User object.

The go back button switches back to the main activity and sends the updated user over.