Added subscription lists to users, made IDs more concise
The database for users now has a field subscriptionList. This is an ArrayList of type string that contains the names/identifiers of each experiment that user is subbed to. Storing them as a field makes more sense than storing them as a collection.

When experiments get functionality for subscribing, run user.addSubscription(Experiment e) to have a user subscribe to it. To dereference the list into experiment objects, run a for loop and check the experiments collection for those names.