Fixed a bug with User profiles and adding trials,  and added support for viewing other people's profiles
Prior to this clicking add trial would generate a new UID every time for unknown reasons. I changed it to pass the user so this no longer happens.

I also changed some code in the DisplayProfile fragment so that it takes 2 profiles, yours and the one that you want to look at. If you are not looking at your own profile you cannot update contact info.